requirements.txt
Appium-Python-Client>=2.11.1
black>=23.3.0
flake8>=6.0.0
interrogate>=1.5.0
isort>=5.12.0
mypy>=1.3.0
pre-commit>=3.3.2
pytest>=7.4.0
python-dotenv>=1.0.0
selenium>=4.10.0
webdriver-manager>=3.8.6
